requirements.txt
Flask==2.3.3
easyocr==1.7.0
google-generativeai==0.3.2
pdf2image==1.16.3
Pillow==10.0.1
numpy==1.24.3
opencv-python-headless==4.8.1.78
torch==2.0.1
torchvision==0.15.2
Werkzeug==2.3.7
gunicorn==21.2.0